+++ src/provider/taskProvider.ts
@@ -4,7 +4,6 @@ import * as vscode from 'vscode';
 import { extensionPath } from '../extension';
 import {
 	getLanguage,
-	pathExists,
 	readJsonFile,
 	replaceBackslashes,
 	writeJsonFile,
@@ -15,6 +14,7 @@ import {
 	JsonInnerTask,
 	JsonTask,
 	Languages,
+	OperatingSystems,
 	Task,
 	Tasks,
 } from '../utils/types';
@@ -134,15 +134,18 @@ export class TaskProvider implements vscode.TaskProvider {
     taskJson.command = settings.makePath;
     taskJson.args[1] = `--file=${this._makefileFile}`;
 
-    const isCleanTask = taskJson.label.includes(Tasks.clean);
     const isRunTask = taskJson.label.includes(Tasks.run);
 
     // Makefile arguments that hold single values
     taskJson.args.push(`COMPILATION_MODE=${this.buildMode}`);
-    taskJson.args.push(`EXECUTABLE_NAME=out${this.buildMode}`);
+    if (this._settingsProvider.operatingSystem === OperatingSystems.windows) {
+      taskJson.args.push(`EXECUTABLE_NAME=out${this.buildMode}.exe`);
+    } else {
+      taskJson.args.push(`EXECUTABLE_NAME=out${this.buildMode}`);
+    }
     taskJson.args.push(`LANGUAGE_MODE=${language}`);
 
-    if (!isCleanTask && !isRunTask) {
+    if (!isRunTask) {
       if (language === Languages.c) {
         taskJson.args.push(`C_COMPILER=${settings.cCompilerPath}`);
         if (
@@ -298,35 +301,6 @@ export class TaskProvider implements vscode.TaskProvider {
     );
 
     this.tasks.push(task);
-  }
-
-  public async runDebugTask() {
-    if (!this._activeFolder) return;
-    if (!this.workspaceFolder) return;
-
-    const uriWorkspaceFolder = vscode.Uri.file(this.workspaceFolder);
-    const folder = vscode.workspace.getWorkspaceFolder(uriWorkspaceFolder);
-    const launchPath = path.join(
-      this.workspaceFolder,
-      '.vscode',
-      'launch.json',
-    );
-
-    const configJson: JsonConfiguration | undefined = readJsonFile(launchPath);
-
-    if (!configJson) return;
-
-    const configIdx = getLaunchConfigIndex(configJson, CONFIG_NAME);
-
-    if (configIdx === undefined) return;
-
-    const buildPath = path.join(this._activeFolder, 'build', this.buildMode);
-    if (!pathExists(buildPath)) return;
-
-    await vscode.debug.startDebugging(
-      folder,
-      configJson.configurations[configIdx],
-    );
   }
 
   public get buildMode() {